requirements.txt
django==5.1.7
django-allauth==65.6.0
psycopg2==2.9.10
python-dotenv==1.1.0
django-mdeditor==0.1.20
markdown==3.7